Inline emailFor and phoneFor, differs by usage
