+++ controllers/apply/awards-for-all/confirmation.js
@@ -2,7 +2,28 @@
 const { get } = require('lodash/fp');
 
 const { MIN_START_DATE } = require('./constants');
-const { getEmailFor, getPhoneFor } = require('../../../common/contacts');
+
+function getEmailFor(country) {
+    const countryEmail = {
+        'default': 'general.enquiries@tnlcommunityfund.org.uk',
+        'england': 'afe@tnlcommunityfund.org.uk',
+        'scotland': 'advicescotland@tnlcommunityfund.org.uk',
+        'northern-ireland': 'enquiries.ni@tnlcommunityfund.org.uk',
+        'wales': 'wales@tnlcommunityfund.org.uk'
+    }[country];
+
+    return countryEmail || 'general.enquiries@tnlcommunityfund.org.uk';
+}
+
+function getPhoneFor(country) {
+    const countryPhone = {
+        'scotland': '0300 123 7110',
+        'northern-ireland': '028 9055 1455',
+        'wales': '0300 123 0735'
+    }[country];
+
+    return countryPhone || '0345 4 10 20 30';
+}
 
 module.exports = function({ locale, data = {} }) {
     const localise = get(locale);
